Repackage to com.github.faarmis
this repackage use to indicate that this library is official reviewed and safe to use by FAARMIS team

This commit also...
- Add NOTICE file
- Update README.md
