+++ NOTICE
@@ -1,0 +1,7 @@
+Android Signature Pad
+Copyright 2021 FAARMis
+
+The Initial developer of main part of library, which are copied from, derived from
+or inspired by Android Signature Pad (https://github.com/gcacace/android-signaturepad),
+by Gianluca Cacace (https://github.com/gcacace)
+Copyright 2014-2016 Gianluca Cacace, All Rights Reserved.
