+++ signature-pad/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.github.gcacace.signaturepad"
+          package="com.github.faarmis.signaturepad"
           android:versionCode="110"
           android:versionName="1.1.0">
